requirements.txt
appnope==0.1.2
asgiref==3.3.4
asttokens==2.0.5
async-generator==1.10
attrs==21.4.0
autopep8==1.5.7
backcall==0.2.0
backports.functools-lru-cache==1.6.4
beautifulsoup4==4.10.0
braillelib==0.0.1
Brotli==1.0.9
brotlipy==0.7.0
certifi==2022.9.24
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.0.12
ChatterBot==1.0.4
chatterbot-corpus==1.2.0
click==8.1.3
contourpy==1.0.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
cycler==0.11.0
dash==2.7.0
dash-bootstrap-components==1.2.1
dash-colorscales==0.0.4
dash-core-components==2.0.0
dash-daq==0.5.0
dash-html-components==2.0.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
dj-rest-auth==2.1.7
Django==3.2.4
django-allauth==0.44.0
django-cors-headers==3.7.0
django-phonenumber-field==6.0.0
django-rest-auth==0.9.5
django-rest-knox==4.2.0
django-sslserver==0.22
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
djoser==2.1.0
drfpasswordless==1.5.7
dtale==2.9.0
entrypoints==0.4
et-xmlfile==1.0.1
executing==0.8.3
Flask==2.2.2
Flask-Compress==1.13
fonttools==4.38.0
future==0.18.2
greenlet==1.1.2
h11==0.13.0
idna==2.10
importlib-metadata==5.1.0
ipykernel==6.9.1
ipython==8.1.1
itsdangerous==2.1.2
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jupyter-client==7.1.2
jupyter-core==4.9.2
kiwisolver==1.4.4
lz4==4.0.2
MarkupSafe==2.1.1
mathparse==0.1.2
matplotlib==3.6.2
matplotlib-inline==0.1.3
missingno==0.4.2
munkres==1.1.4
nest-asyncio==1.5.4
networkx==2.8.8
nltk==3.7
numpy==1.23.5
oauthlib==3.1.1
openpyxl==3.0.10
outcome==1.1.0
packaging==21.3
pandas==1.5.2
parso==0.8.3
patsy==0.5.3
pdf2image==1.16.0
pexpect==4.8.0
phonenumbers==8.12.45
pickleshare==0.7.5
Pillow==9.2.0
Pint==0.19.1
pip==21.1.2
playwright==1.19.1
plotly==5.11.0
prompt-toolkit==3.0.27
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pybraille==1.0.0
pycodestyle==2.7.0
pycparser==2.20
pyee==8.1.0
Pygments==2.11.2
PyJWT==2.1.0
pymongo==3.12.3
pyOpenSSL==20.0.1
pyotp==2.6.0
pyparsing==3.0.9
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.7.5
python-decouple==3.6
python3-openid==3.2.0
pytz==2021.1
PyYAML==3.13
pyzmq==22.3.0
regex==2022.3.15
requests==2.27.1
requests-oauthlib==1.3.0
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.1
selenium==4.1.3
setuptools==49.6.0.post20210108
six==1.16.0
sniffio==1.2.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sortedcontainers==2.4.0
soupsieve==2.3.1
SQLAlchemy==1.2.19
sqlparse==0.4.1
squarify==0.4.3
stack-data==0.2.0
statsmodels==0.13.5
strsimpy==0.1.9
tenacity==8.1.0
threadpoolctl==3.1.0
toml==0.10.2
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
trio==0.20.0
trio-websocket==0.9.2
twilio==7.8.0
typing-extensions==3.10.0.0
unicodedata2==15.0.0
uritemplate==3.0.1
urllib3==1.26.5
wcwidth==0.2.5
websockets==10.1
Werkzeug==2.2.2
wheel==0.36.2
wsproto==1.1.0
xarray==2022.11.0
xlrd==2.0.1
zipp==3.11.0
